Remove DQM version
